app/realtime/requirements.txt
fastapi==0.115.9
uvicorn==0.35.0
pandas==2.2.3
plotly==6.1.2
python-dotenv==1.1.0
jinja2==3.1.6
supabase==2.15.1